+++ demo/sitetools/core.cljs
@@ -2,7 +2,7 @@
   (:require [clojure.string :as string]
             [goog.events :as evt]
             [goog.history.EventType :as hevt]
-            [reagent.core :as reagent :refer [atom partial]]
+            [reagent.core :as r]
             [reagent.debug :refer-macros [dbg log dev?]]
             [reagent.interop :as i :refer-macros [.' .!]])
   (:import [goog History]
@@ -17,19 +17,19 @@
 
 ;;; Configuration
 
-(defonce config (atom {:page-map {"index.html" [:div "Empty"]}
-                       :page-titles {}
-                       :body [page-content]
-                       :site-dir "outsite/public"
-                       :css-infiles ["site/public/css/main.css"]
-                       :css-file "css/built.css"
-                       :js-file "js/main.js"
-                       :js-dir "js/out"
-                       :main-div "main-content"
-                       :allow-html5-history false}))
+(defonce config (r/atom {:page-map {"index.html" [:div "Empty"]}
+                         :page-titles {}
+                         :body [page-content]
+                         :site-dir "outsite/public"
+                         :css-infiles ["site/public/css/main.css"]
+                         :css-file "css/built.css"
+                         :js-file "js/main.js"
+                         :js-dir "js/out"
+                         :main-div "main-content"
+                         :allow-html5-history false}))
 
-(defonce page (atom "index.html"))
-(defonce page-state (atom {:has-history false}))
+(defonce page (r/atom "index.html"))
+(defonce page-state (r/atom {:has-history false}))
 
 (defn register-page
   ([pageurl comp]
@@ -60,7 +60,7 @@
                       (fn [e]
                         (.preventDefault e)
                         (reset! page p)
-                        (reagent/next-tick
+                        (r/next-tick
                          #(set! (.-scrollTop (.-body js/document))
                                 0)))
                       identity))
@@ -84,7 +84,7 @@
 
 (add-watch page ::title-watch
            (fn [_ _ _ p]
-             (when reagent/is-client
+             (when r/is-client
                (set! (.-title js/document) (get-title)))
              ;; First title on a page wins
              #_(reset! page-title "")))
@@ -92,7 +92,7 @@
 ;;; History
 
 (defn use-html5-history []
-  (when reagent/is-client
+  (when r/is-client
     (let [proto (.' js/window :location.protocol)]
       (and (:allow-html5-history @config)
            (.isSupported Html5History)
@@ -125,7 +125,7 @@
                       (reset! page (if (and tb (== 0 (.indexOf t tb)))
                                      (subs t (count tb))
                                      t)))
-                    (reagent/flush)))
+                    (r/flush)))
       (add-watch page ::history
                  (fn [_ _ oldp newp]
                    (when-not (= oldp newp)
@@ -167,7 +167,7 @@
         main (str (:js-file c) timestamp)
         css-file (:css-file c)
         opt-none (:opt-none c)]
-    (reagent/render-to-static-markup
+    (r/render-to-static-markup
      [:html
       [:head
        [:meta {:charset "utf-8"}]
@@ -186,7 +186,7 @@
 
 (defn gen-page [page-name timestamp]
   (reset! page page-name)
-  (let [b (reagent/render-component-to-string (body))]
+  (let [b (r/render-component-to-string (body))]
     (str "<!doctype html>"
          (html-template {:title (get-title)
                          :body b
@@ -234,23 +234,16 @@
   (log "Generating site")
   (swap! config merge (js->clj opts :keywordize-keys true))
   (let [dir (:site-dir @config)
-        written (atom #{})
-        timestamp (str "?" (.' js/Date now))
-        one-page (fn [] (first (filter
-                                (fn [x] (nil? (@written x)))
-                                (keys (:page-map @config)))))]
-    (loop [f (one-page)]
-      (when f
-        (swap! written conj f)
-        (write-file (path-join dir f)
-                    (gen-page f timestamp))
-        (recur (one-page))))
+        timestamp (str "?" (.' js/Date now))]
+    (doseq [f (keys (:page-map @config))]
+      (write-file (path-join dir f)
+                  (gen-page f timestamp)))
     (write-resources dir))
   (log "Wrote site"))
 
 (defn start! [site-config]
   (swap! config merge site-config)
-  (when reagent/is-client
+  (when r/is-client
     (let [conf (when (exists? js/pageConfig)
                  (js->clj js/pageConfig :keywordize-keys true))
           page-name (:page-name conf)]
@@ -260,6 +253,6 @@
           (set-start-page page-name))
         (setup-history page-name)
         (set! (.-title js/document) (get-title)))
-      (reagent/render-component (body)
-                                (.' js/document getElementById
-                                    (:main-div @config))))))
+      (r/render-component (body)
+                          (.' js/document getElementById
+                              (:main-div @config))))))