+++ examples/todomvc/devsrc/todomvc/dev.cljs
@@ -1,6 +1,0 @@
-(ns todomvc.dev
-  (:require [todomvc.core :as example]
-            [figwheel.client :as fw]))
-
-(fw/start {:on-jsload example/run
-           :websocket-url "ws://localhost:3449/figwheel-ws"})